Creación de vista cliente y agregado el estilo para abonos de una venta fiada

+++ src/controllers/controller.py
@@ -1,5 +1,4 @@
-from flask import render_template, request, redirect, url_for
-from flask import session
+from flask import render_template, request, redirect, url_for, jsonify
 from src.models.model import *
 from datetime import datetime, timedelta
 # from src.db import db
@@ -77,10 +76,10 @@ def index():
     else: 
         return redirect(url_for('login'))
     
-def productos():
+def productos(msg):
     productos = session.query(Producto).all()
     if obj.get_boolean() is True:
-        return render_template('productos.html', productos=productos)
+        return render_template('productos.html', productos=productos, msg=msg)
     else: 
         return redirect(url_for('login'))
     
@@ -130,7 +129,12 @@ def modificar_productos(idProducto):
     return redirect(url_for('productos'))
 
 def eliminar_productos(idProducto):
-    producto = session.query(Producto).get(idProducto)
-    session.delete(producto)
-    session.commit()
-    return redirect(url_for('productos'))
+    productos_ventas = session.query(ProductosVentas).filter_by(idProducto=idProducto).all()
+    if(len(productos_ventas)==0): 
+        producto = session.query(Producto).get(idProducto)
+        session.delete(producto)
+        session.commit()
+        return productos("eliminado")
+    else:
+        return productos("imposible")
+
